% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{c.elco}
\alias{c.elco}
\title{Combine \code{elco} objects into a vector or list}
\usage{
\method{c}{elco}(...)
}
\arguments{
\item{...}{objects to be concatenated.  All \code{\link[base]{NULL}} entries
    are dropped before method dispatch unless at the very beginning of the
    argument list.}
}
\value{
\code{...} concatenated into one vector.
}
\description{
S3 method for \code{elco} objects (see \code{\link[=c]{c()}}).
}
\examples{
c(elco::chno$C[[1]], elco::chno$C[-1])
c(elco::chno$C[[1]], elco::chno$N[-1])

}
