% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chno}
\alias{chno}
\title{Simulated C, H, N, O contents, and sample masses for five samples}
\format{
A data frame with 5 rows and 5 columns:
\describe{
\item{C}{A numeric vector with the relative mass fraction of C [g/g].}
\item{H}{A numeric vector with the relative mass fraction of H [g/g].}
\item{N}{A numeric vector with the relative mass fraction of N [g/g].}
\item{O}{A numeric vector with the relative mass fraction of O [g/g].}
\item{sample_mass}{A numeric vector with the mass of the samples [mg].}
}
}
\usage{
chno
}
\description{
A \code{data.frame} with simulated C, H, N, O contents, and sample masses for five
samples.
}
\keyword{datasets}
