% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{el_units_allowed}
\alias{el_units_allowed}
\title{Allowed units for \code{elco} objects}
\format{
A data frame with 96 rows and 2 columns:
\describe{
\item{type}{A character value classifying measurement units by their
nominator and denominator.}
\item{unit_symbol}{A character value representing a measurement unit.}
}
}
\usage{
el_units_allowed
}
\description{
A data frame listing allowed units for \code{elco} objects.
}
\keyword{datasets}
