% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco-extract}
\alias{elco-extract}
\alias{[.elco}
\alias{[[.elco}
\alias{[<-.elco}
\alias{[[<-.elco}
\title{Extract or replace parts of an \code{elco} object}
\usage{
\method{[}{elco}(x, i)

\method{[[}{elco}(x, i)

\method{[}{elco}(x, i, j, ...) <- value

\method{[[}{elco}(x, i, j, ...) <- value
}
\arguments{
\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
Extract or replace parts of an \code{elco} object
}
\examples{
elco::chno$C[1]

elco::chno$C[[1]]

x <- elco::chno$C
y <- elco::chno$N

# returns `elco` object because both have the same el_symbol attribute
x[1] <- x[2]
class(x)

# strips off the elco class because the el_symbol attribute differs
y[1] <- x[2]
class(y)

x <- elco::chno$C
y <- elco::chno$N

# returns `elco` object because both have the same el_symbol attribute
x[[1]] <- x[2]
class(x)

# strips off the elco class because the el_symbol attribute differs
y[[1]] <- x[2]
class(y)

}
