% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_groupGeneric.R
\name{elco-groupGeneric}
\alias{elco-groupGeneric}
\alias{Math.elco}
\alias{Ops.elco}
\title{S3 Group Generic Functions for objects of class \code{elco}}
\usage{
\method{Math}{elco}(x, ...)

\method{Ops}{elco}(e1, e2)
}
\arguments{
\item{x}{An object of class \code{elco}.}

\item{...}{further arguments passed to methods.}

\item{e1}{An object of class \code{elco}.}

\item{e2}{An object of class \code{elco} which is
added to/subtracted from/multiplied with/used for division of \code{e1}.}
}
\description{
\code{Math}, \code{Ops} and \code{Summary} group generic methods for \code{elco} objects (see
\code{\link[base:groupGeneric]{base::groupGeneric()}} for a comprehensive list of available methods).
}
\examples{
# Math
abs(elco::chno$C)

x <- elco::chno

x$C + x$C

# drops elco class because these are two different element symbols:
x$C + x$N

# drops elco class because the operation is not meaningful for chemical
# elements:
x$C * x$C

}
