% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco-doc.R, R/elco-vctrs.R
\name{elco-vctrs}
\alias{elco-vctrs}
\alias{vec_restore.elco}
\alias{vec_proxy.elco}
\alias{vec_ptype2.elco.elco}
\alias{vec_cast.elco.elco}
\title{Internal vctrs methods}
\usage{
\method{vec_restore}{elco}(x, to, ...)

\method{vec_proxy}{elco}(x, ...)

\method{vec_ptype2}{elco.elco}(x, y, ...)

\method{vec_cast}{elco.elco}(x, to, ...)
}
\description{
Internal vctrs methods
}
\examples{
# vec_restore.elco
x <- elco::chno$C
vctrs::vec_restore(vctrs::vec_data(x), to = x)

# vec_proxy.elco
vctrs::vec_proxy(elco::chno$C)

# vec_ptype2.elco.elco
vctrs::vec_ptype2(x = elco::chno$C, y = elco::chno$C)

# vec_cast.elco.elco
vctrs::vec_cast(x = elco::chno$C, to = elco::chno$C)
vctrs::vec_cast(x = elco::chno$C, to = elco::chno$N)

}
\keyword{internal}
