% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco_check_elco}
\alias{elco_check_elco}
\title{Checks if an object is of class \code{elco}.}
\usage{
elco_check_elco(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{elco}.
}
\description{
\code{elco_check_elco} checks if an object is of class \code{\link[=elco_new_elco]{elco()}}.
}
\keyword{internal}
