% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irms.R
\name{elco_check_irms}
\alias{elco_check_irms}
\title{Checks if an object is of class \code{irms}.}
\usage{
elco_check_irms(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{\link[=elco_new_irms]{irms()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{elco_check_irms} checks if an object is of class \code{\link[=elco_new_irms]{irms()}}.
}
\keyword{internal}
