% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irms_std.R
\name{elco_check_irms_std}
\alias{elco_check_irms_std}
\title{Checks if an object is of class \code{irms_std}.}
\usage{
elco_check_irms_std(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{\link[=elco_new_irms_std]{irms_std()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{elco_check_irms_std} checks if an object is of class \code{\link[=elco_new_irms_std]{irms_std()}}.
}
\keyword{internal}
