% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco_check_unit_conversion}
\alias{elco_check_unit_conversion}
\title{Checks if a unit conversion is valid.}
\usage{
elco_check_unit_conversion(x, to, silent = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link[units:units]{units}}.}

\item{to}{A character value representing the symbol for a unit into which
\code{x} should be converted.}
}
\value{
A logical value indicating if the conversion is valid (\code{TRUE})
or not (\code{FALSE}).
}
\description{
\code{elco_check_unit_conversion} takes an object of class \code{\link[units:units]{units}}
and a character value representing a unit it should be converted into and checks if the
conversion is valid.
}
