% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{elco_check_xrf}
\alias{elco_check_xrf}
\title{Checks if an object is of class \code{xrf}.}
\usage{
elco_check_xrf(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{\link[=elco_new_xrf]{xrf()}}.
}
\description{
\code{elco_check_xrf} checks if an object is of class \code{\link[=elco_new_xrf]{xrf()}}.
}
\keyword{internal}
