% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_el_symbol.R
\name{elco_el_symbol}
\alias{elco_el_symbol}
\alias{elco_el_symbol.elco}
\alias{elco_el_symbol<-}
\alias{elco_el_symbol<-.elco}
\alias{elco_el_symbol<-.quantities}
\alias{elco_set_el_symbol}
\alias{elco_set_el_symbol.elco}
\alias{elco_set_el_symbol.quantities}
\title{Getting or setting chemical element information}
\usage{
elco_el_symbol(x)

\method{elco_el_symbol}{elco}(x)

elco_el_symbol(x) <- value

\method{elco_el_symbol}{elco}(x) <- value

\method{elco_el_symbol}{quantities}(x) <- value

elco_set_el_symbol(x, value)

\method{elco_set_el_symbol}{elco}(x, value)

\method{elco_set_el_symbol}{quantities}(x, value)
}
\arguments{
\item{x}{An object.}

\item{value}{A character value representing a chemical element symbol with
which to set or replace an existing attribute \code{"el_symbol"} in \code{x}.}
}
\value{
\code{x} converted to an \code{\link[=elco_new_elco]{elco}} object.

\code{x} with attribute value for \code{el_symbol} replaced by \code{value}.
}
\description{
Getting or setting chemical element information
}
\examples{
# getting element information from an `elco` object
elco_el_symbol(elco::chno$C)

## setting element information in an `elco` object
x <- elco::chno$C
elco_el_symbol(x) <- "C"

# pipe-friendly
elco::chno$C \%>\% elco_set_el_symbol("C")

## setting element information in a `quantities` object
x <- elco_drop_elco(elco::chno$C)

elco_el_symbol(x) <- "C"

# pipe-friendly
x <-
  elco_drop_elco(elco::chno$C) \%>\%
  elco_set_el_symbol("C")

}
