% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco_identify_elco_df}
\alias{elco_identify_elco_df}
\title{Identifies columns of class \code{elco} in a \code{data.frame}.}
\usage{
elco_identify_elco_df(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
A logical vector with an element for each column in \code{x} indicating if
the column is of class \code{elco} or not.
}
\description{
\code{elco_identify_elco_df} identifies columns in a \code{data.frame} that
are of class \code{\link[elco_elco_new_elco]{elco}}.
}
