% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_irms_extract_standards.R
\name{elco_irms_check_standards}
\alias{elco_irms_check_standards}
\title{Takes an object of class \code{irms} and checks if it contains only measurements on standards.}
\usage{
elco_irms_check_standards(x)
}
\arguments{
\item{x}{An object of class \code{\link[=elco_new_irms]{irms()}}}
}
\value{
\code{x}.
}
\description{
\code{elco_irms_check_standards} takes an object of class \code{irms} and checks if it has only rows
referring to standards in \code{irms_standards}.
}
