% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_irms_correct_elements.R
\name{elco_irms_correct_elements}
\alias{elco_irms_correct_elements}
\title{Heuristic function to correct IRMS element contents by computing a manual calibration function from the standard measurements and apply this to predict element contents in samples.}
\usage{
elco_irms_correct_elements(
  x,
  element = "C",
  irms_standards_to_use = elco::irms_standards,
  standards = irms_standards$standard_name,
  by_file = TRUE,
  plotit = FALSE
)
}
\arguments{
\item{x}{An object of class \code{\link[=elco_new_irms]{irms()}}.}

\item{element}{A character value representing the chemical element for which
to correct the mass fraction values. This must be one of "C" or "N".}

\item{irms_standards_to_use}{A \code{data.frame} with the same format as
\code{\link[=irms_standards]{irms_standards()}}. This can optionally be used to make available other
reference values for plotting.}

\item{standards}{A character vale specifying standards to use for computing
the regression equation. This must be one of
\code{irms_standard::irms_standard$standard_name} or, if specified, values in the
same column of \code{irms_standards_to_use}. Default is to use all
standards.}

\item{by_file}{A logical value indicating if medians of standards are
computed across different files as indicated by \code{x$file_id} or (\code{FALSE})
individually for each file (\code{TRUE}).}

\item{plotit}{A logical value indicating if a plot for checking should be
printed (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
\code{x} with corrected element content.
}
\description{
\code{elco_irms_correct_elements} is a heuristic function to correct measured element contents of samples
during IRMS analysis. Correction can be performed for either C or N and, if multiple
files are batch-processed, using either standards from all files or only from the file where the respective
sample is located in.
The procedure is as follows:
\enumerate{
\item For each measured standard, assign the known reference element mass content from
\code{\link{irms_standards}}.
\item Use the information on the mass of the measured samples and the assigned element mass content values
to compute the respective absolute known element content [mg].
\item For all measured standards, compute a linear regression model that predicts the absolute
element content in dependency of the respective signal area.
\item For each sample, use the regression model and the respective measured signal area to compute
the absolute element content.
\item Divide the absolute predicted element contents by the samples' masses [mg] to get the mass fraction of
the element.
\item (Optionally): create plot: (1) showing uncorrected and corrected element mass fractions
for all measured standards and samples. This plot may be used as a visual check of the heuristic
correction procedure and shows how the values for all samples are corrected. In addition, for each
fitted regression model, the measured absolute element masses are plotted in dependency of the signal
areas with the regression line. These plot can be used to assess if the regression model captures well
the measured data for the standards.
}
}
