% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_irms_correct_isotopes.R
\name{elco_irms_correct_isotopes}
\alias{elco_irms_correct_isotopes}
\title{Heuristic function to correct IRMS isotope signals by substracting differences measured values for standards and their known isotope signature.}
\usage{
elco_irms_correct_isotopes(
  x,
  isotope = "13C",
  t = 5,
  by_file = TRUE,
  plotit = FALSE
)
}
\arguments{
\item{x}{An object of class \code{\link[elco:elco_new_irms]{irms}}.}

\item{isotope}{A character value representing the isotope for which to correct isotope signatures.
This must be one of "13C" or "15N".}

\item{t}{An integer value specifying the number of measurements that must be available for a particular
standard to make it eligble for being a reference for correcting the measured isotope values.}

\item{by_file}{A logical value indicating if medians of standards are computed across different files
as indicated by \code{x$file_id} or (\code{FALSE}) individually for each file (\code{TRUE}).}

\item{plotit}{A logical value indicating if a plot for checking should be printed (\code{TRUE}) or not
(\code{FALSE}).}
}
\value{
nothig.
}
\description{
\code{elco_irms_correct_isotopes} is a heuristic function to correct measured isotope signatures of samples
during IRMS analysis. Correction can be performed for either \eqn{^{13}}C or \eqn{^{13}}N and, if multiple
files are batch-processed, using either standards from all files or only from the file where the respective
sample is located in.
The procedure is as follows:
\enumerate{
  \item For each standard, compute the median isotope signature and median signal area for the respective isotope.
  The median is used instead of an
  average to prevent potential outliers from having too large an effect.
  \item For each standard, record the number of measurements. This is used to exclude using standards
  as reference with too few measurmements since these will have large uncertainties in their estimated
  median. From the list of standards, discard all with <t measurements, where t is an integer value >1.
  \item For each sample (including the measured standards), compute the absolute difference in the signal area for the respective isotope to all
  standards' median signal area values. Record for each sample the standard with the least absolute difference.
  \item Look up the known reference isotope signatures for each assigned standard and substract this value
  from the measured value.
  \item (Optionally): create a plot showing uncorrected and corrected isotope signatures for all measured
  standards along their known reference values. This plot may be used as a visual check of the heuristic
  correction procedure.
}
}
\note{
___ todo: Klaus suggested that to remove outliers in standard measurements that may not be discarded
by the median approach, we should define a threshold value for each standard to exclude outliers before
computing medians. Check if this should still be done.
}
