% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_irms_import_csv.R
\name{elco_irms_import_csv}
\alias{elco_irms_import_csv}
\title{Reformats raw csv element analyzer isotope ratio mass spectrometry (EA-IRMS) data as exported by the vendor software.}
\usage{
elco_irms_import_csv(files)
}
\arguments{
\item{files}{A vector of character values representing paths to csv files with EA-IRMS raw data as exported
by the vendor software of the IRMS.}
}
\value{
An object of class \code{\link[=elco_new_irms]{irms}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{elco_irms_import_csv} reads one or more csv files that contain raw element analyzer isotope ratio mass spectrometry
(EA-IRMS) data as exported by the vendor software of the IRMS spectrometer and tidies the
data. This includes reformatting columns and rows, deleting unneeded columns and rows,
adding units, and renaming columns.
}
\keyword{internal}
