% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_irms_write_standards.R
\name{elco_irms_write_standards}
\alias{elco_irms_write_standards}
\title{Takes an object of class \code{irms_std} and appends it to an existing rds file.}
\usage{
elco_irms_write_standards(x, file, initialize = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link[=elco_new_irms_std]{irms_std}}.}

\item{file}{A character value representing a path to an existing rds file with
an object of class \code{\link[=elco_new_irms_std]{irms_std}} or to such a file
to be created.}

\item{initialize}{A logical value indicating if the file pointed to by \code{file}
already exists and \code{x} should be appended to this (\code{TRUE}), or does not exist
and a new file should be created (\code{TRUE}).}

\item{verbose}{A logical value indicating if messages should be printed (\code{TRUE}) or
not (\code{FALSE}).}
}
\value{
nothing.
}
\description{
\code{elco_irms_extract_standards} takes an object of class \code{irms_std} and extracts all rows
referring to standards in \code{irms_standards}.
}
