% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco_new_elco}
\alias{elco_new_elco}
\title{Creates an object of class \code{elco}}
\usage{
elco_new_elco(x, el_symbol)
}
\arguments{
\item{x}{A numeric vector that is an object of class
\code{\link[quantities:quantities]{quantities}} with each element representing an
element content.}

\item{el_symbol}{A character value representing the symbol for a chemical element.}
}
\value{
An object of class \code{elco}. This is identical to \code{x}, but has
an additional attribute \code{"el_symbol"} and class attribute.
}
\description{
\code{elco_new_elco} is the constructor function for objects of class
\code{elco}. An object of class \code{elco} is a numeric vector with each element
representing a chemical element content value. At the same time, the numeric vector
is an object of class \code{\link[quantities:quantities]{quantities}} with an
additional attribute \code{"el_symbol"} with the respective element symbol.
}
