% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_nosc.R
\name{elco_nosc}
\alias{elco_nosc}
\alias{elco_or}
\alias{elco_du}
\title{Carbon oxidation state, carbon oxidative ratio, and degree of unsaturation}
\usage{
elco_nosc(C, H, N, O)

elco_or(C, H, N, O)

elco_du(C, H, N)
}
\arguments{
\item{C}{A quantities object [mol_C].}

\item{H}{A quantities object [mol_H].}

\item{N}{A quantities object [mol_N].}

\item{O}{A quantities object [mol_O].}
}
\value{
\itemize{
\item \code{elco_nosc}: A \code{\link[quantities:quantities]{quantities}} object with the carbon
oxidation state.
}

\itemize{
\item \code{elco_or}: A \code{\link[quantities:quantities]{quantities}} object with the
oxidative ratio.
}

\itemize{
\item \code{elco_du}: A \code{\link[quantities:quantities]{quantities}} object with the degree
of unsaturation.
}
}
\description{
Functions to compute the carbon oxidation state, carbon oxidative ratio, and
degree of unsaturation with \code{\link[quantities:quantities]{quantities}} objects.
}
\section{Functions}{
\itemize{
\item \code{elco_nosc()}: Computes the carbon oxidation
state using the elemental composition of C, H, N, O.

\item \code{elco_or()}: Computes the carbon oxidation
state using the elemental composition of C, H, N, O.

\item \code{elco_du()}: Computes the carbon degree
of unsaturation using the elemental composition of C, H, N.

}}
\examples{
# sample data with elemental contents in mol
d <-
  elco::chno \%>\%
  dplyr::mutate(
    dplyr::across(
      ! dplyr::all_of("sample_mass"),
      function(.x) elco_convert(.x * sample_mass, to = "mol")
    )
  )

## NOSC
d \%>\%
  dplyr::mutate(nosc = elco_nosc(C = C, H = H, N = N, O = O))

## oxidative ratio
d \%>\%
  dplyr::mutate(or = elco_or(C = C, H = H, N = N, O = O))

## degree of unsaturation
d \%>\%
  dplyr::mutate(du = elco_du(C = C, H = H, N = N))

}
