% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elco_reclass_elco}
\alias{elco_reclass_elco}
\title{Restores the class attribute of an object with the attributes of an object of class \code{elco}}
\source{
The function is modified from \code{quantities:::reclass}
(\url{https://github.com/r-quantities/quantities/blob/master/R/utils.R}).
}
\usage{
elco_reclass_elco(x)
}
\arguments{
\item{x}{A numeric vector with the attributes of an object of class
\code{\link[elco:elco_new_elco]{elco}}.}
}
\value{
An object of class \code{elco}.
}
\description{
\code{elco_reclass_elco} restores the class attribute of an object with the attributes
(\code{units}, \code{errors}, \code{el_symbol}) of an object of class \code{elco}.
}
