% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elco_units}
\alias{elco_units}
\title{Custom units defined by the 'elco' package}
\format{
A data frame with 117 rows and 6 columns:
\describe{
\item{thing}{A character value defining the chemical element or other
entity for which a unit is defined.}
\item{mol}{A character value defining the unit for molar amount of the
entity. This is \code{NA} if no such unit is defined.}
\item{mol_names}{A character value defining the unit names corresponding to
column \code{mol}.}
\item{g}{A character value defining the unit for mass of the
entity. This is \code{NA} if no such unit is defined.}
\item{g_names}{A character value defining the unit names corresponding to
column \code{g}.}
\item{g_to_mol}{A character value defining the conversion constants between
the mol and g units.}
}
}
\usage{
elco_units
}
\description{
A data frame listing the custom units defined by the 'elco' package
}
\keyword{datasets}
