% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_xrf_export_csv.R
\name{elco_xrf_export_csv}
\alias{elco_xrf_export_csv}
\title{Reformats raw csv X-ray fluorescence data as exported by the vendor software.}
\usage{
elco_xrf_export_csv(x, file)
}
\arguments{
\item{x}{An object of class \code{\link[=elco_new_xrf]{xrf}}.}

\item{file}{A character value representing a path and file name (csv file) where to export
the reformatted data to. The two exported csv files are named using the value of \code{file}:
\enumerate{
\item A file named with the value of \code{file}.
\item A file named with the value of \code{file} and with additional \code{"_metadata"}, containing
the metadata (variable type and unit).
}}
}
\value{
nothing.
}
\description{
\code{elco_xrf_export_csv} takes an object of class xrf and exports two csv files:
one with the measured data and one with metadata indicating data types and units of measurement.
}
