% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.xrf}
\alias{plot.xrf}
\title{Plots an object of class xrf.}
\usage{
\method{plot}{xrf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=elco_new_xrf]{xrf()}}.}

\item{...}{Further arguments, will be ignored}
}
\value{
An object of class \code{\link[ggplot2:ggplot]{ggplot2()}}
}
\description{
\code{plot.xrf} is the plot method for objects of class
\code{xrf}. This function can be used to create plots for checking the
data.
}
