% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{print.elco}
\alias{print.elco}
\title{Prints an object of class \code{elco}.}
\usage{
\method{print}{elco}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[elco:elco_new_elco]{elco}}.}

\item{...}{Additional arguments, will be ignored.}
}
\value{
\code{x}.
}
\description{
Prints an object of class \code{elco}.
}
