% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xrf_calibration}
\alias{xrf_calibration}
\title{Correction factors for XRF element contents for samples with differing masses.}
\format{
A data frame with 60 rows and 3 columns:
\describe{
  \item{el_symbol}{A character vector with the symbols of chemical elements for which
  correction factors were computed.}
  \item{sample_mass}{A character vector with discrete sample mass values [mg] for
  which the correction factors were computed.}
  \item{correction_factor}{A \code{quantities} vector with the mean, standard deviation, and
  unit for each correction factor. See \code{\link[elco:elco_xrf_correct_elements]{elco_xrf_correct_elements}}}
  for how these values are used.
}
}
\usage{
xrf_calibration
}
\description{
A \code{data.frame} with chemical element-specific heuristic correction factors for XRF element contents (relative to
a reference mass of 500 mg) measured with masses smaller than the reference mass.
}
\keyword{datasets}
