% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco-extract}
\alias{elco-extract}
\alias{[.elco}
\alias{[[.elco}
\alias{[<-.elco}
\alias{[[<-.elco}
\title{Extract or replace parts of an \code{elco} object}
\usage{
\method{[}{elco}(x, i)

\method{[[}{elco}(x, i)

\method{[}{elco}(x, i, j, ...) <- value

\method{[[}{elco}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An \code{elco} object to subset values from or where to replace values.}

\item{i, j, ...}{indices specifying elements to extract or replace. See
\code{\link[base:Extract]{base::Extract}}.}

\item{value}{An object of class \code{elco} which replaces elements in \code{x}.}
}
\value{
An object of class \code{elco} with subsetted or replaced values.
}
\description{
Extract or replace parts of an \code{elco} object
}
\examples{
elco::chno$C[1]

elco::chno$C[[1]]

x <- elco::chno$C
y <- elco::chno$N
z <- NA_real_

# returns `elco` object because both have the same el_symbol attribute
x[1] <- x[2]
class(x)

# strips off the elco class because the el_symbol attribute differs
y[1] <- x[2]
class(y)

# strips off the elco class because z is no elco object
x[1] <- z
class(x)

x <- elco::chno$C
y <- elco::chno$N

# returns `elco` object because both have the same el_symbol attribute
x[[1]] <- x[2]
class(x)

# strips off the elco class because the el_symbol attribute differs
y[[1]] <- x[2]
class(y)

}
