% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco-package.R
\docType{package}
\name{elco-package}
\alias{elco}
\alias{elco-package}
\title{elco: Measurement units for chemical elements}
\description{
'elco' extends the 'units' package by defining measurement units for each chemical element. This allows to consider the identity of the chemical element in calculations with units.
}
\author{
\strong{Maintainer}: Henning Teickner \email{h_teic01@uni-muenster.de} (\href{https://orcid.org/0000-0002-3993-1182}{ORCID})

Authors:
\itemize{
  \item Klaus-Holger Knorr \email{kh.knorrr@wwu.de}
}

}
\keyword{internal}
