% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco-package.R
\docType{package}
\name{elco-package}
\alias{elco}
\alias{elco-package}
\title{elco: Handling data on chemical element contents and isotope signatures.}
\description{
elco provides functions to read in raw data on element contents
    from X-ray fluorescence (XRF), gaschromatographic element analysis (EA), and isotope ratio
    mass spectromtery (IRMS), handling these as data.frames with units, and
    computing simple derived variables, such as element ratios.
}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@wwu.de}

Authors:
\itemize{
  \item Klaus-Holger Knorr \email{kh.knorrr@wwu.de}
}

}
\keyword{internal}
