% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irms_std.R
\name{elco_check_irms_std}
\alias{elco_check_irms_std}
\title{Checks if an object is of class \code{irms_std}.}
\usage{
elco_check_irms_std(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{\link[=elco_new_irms_std]{irms_std()}}.
}
\description{
\code{elco_check_irms_std} checks if an object is of class \code{\link[=elco_new_irms_std]{irms_std()}}.
}
\keyword{internal}
