% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_irms_extract_standards.R
\name{elco_irms_extract_standards}
\alias{elco_irms_extract_standards}
\title{Takes an object of class \code{irms} and extracts all rows referring to standards in \code{irms_standards}}
\usage{
elco_irms_extract_standards(x, irms_standards_to_use = elco::irms_standards)
}
\arguments{
\item{x}{An object of class \code{\link[=elco_new_irms]{irms}}}

\item{irms_standards_to_use}{A \code{data.frame} with the same format as
\code{\link[=irms_standards]{irms_standards()}}. This can optionally be used to make available other
reference values for plotting.}
}
\value{
An object of class \code{\link[=elco_new_irms_std]{irms_std}} containing
only the rows in \code{x} where
\code{sample_name \%in\% elco::irms_standards$standard_name}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{elco_irms_extract_standards} takes an object of class \code{irms} and extracts
all rows referring to standards in \code{irms_standards}.
}
\keyword{internal}
