% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_make_units_generic.R
\name{elco_make_units_generic}
\alias{elco_make_units_generic}
\title{Changes units defined by the 'elco' to the respective generic units}
\usage{
elco_make_units_generic(x)
}
\arguments{
\item{x}{A units object in which to change units defined by the 'elco'
to the respective generic units.}
}
\value{
\code{x} with generic units.
}
\description{
For example, \code{g_C} is changed to \code{g} and \code{mol_C} to \code{mol}. The function
does not test whether the result is sensible. E.g. one could have unit
\code{(g_C*g_H)/(g_sample*L)} (see the example) which \code{elco_make_units_generic()}
turns into \code{g/L}.
}
\examples{
# changing g_C into g
units::as_units("g_C", mode = "standard") |>
  elco_make_units_generic()

# changing mmol_C/kg_sample into mol/g
units::as_units("mmol_C/kg_sample", mode = "standard") |>
  elco_make_units_generic()

# make sure to check whether the result makes sense
units::as_units("(g_C * g_H) / (g_sample * L)", mode = "standard") |>
  elco_make_units_generic()


}
