% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\name{elco_new_elco}
\alias{elco_new_elco}
\title{Creates an object of class \code{elco}}
\usage{
elco_new_elco(x, el_symbol)
}
\arguments{
\item{x}{A numeric vector that is an object of class
\code{\link[quantities:quantities]{quantities}} with each element representing an
element content.}

\item{el_symbol}{A character value representing the symbol for a chemical
element.}
}
\value{
An object of class \code{elco}. This is identical to \code{x}, but has
an additional attribute \code{"el_symbol"} and class attribute.
}
\description{
\code{elco_new_elco} is the constructor function for objects of class \code{elco}. An
object of class \code{elco} is a numeric vector with each element representing a
chemical element content value. It is a subclass of the
\code{\link[quantities:quantities]{quantities::quantities()}} class with an additional attribute \code{"el_symbol"}
where the respective element symbol is stored.
}
\examples{
library(quantities)

# store C content value as elco object
elco_new_elco(
  x = quantities::set_quantities(0.42, unit = "mol/g", errors = 0.002),
  el_symbol = "C"
)

}
