% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{elco_new_xrf}
\alias{elco_new_xrf}
\title{Creates an object of class xrf.}
\usage{
elco_new_xrf(x)
}
\arguments{
\item{x}{A \code{data.frame} with a row for each measured sample and the
following columns:
\describe{
\item{sample_id}{A character vector with the sample labels.}
\item{method_name}{A character vector with the names of the measurement methods used.}
\item{time}{A POSIXct vector with the date and time when the sample was measured.}
\item{Na}{A units vector with the measured Na content [\eqn{\mu}g/g].}
\item{Na_flag}{A logical vector indicating if the calibration range was exceeded for \code{Na}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Mg}{A units vector with the measured Mg content [\eqn{\mu}g/g].}
\item{Mg_flag}{A logical vector indicating if the calibration range was exceeded for \code{Mg}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Al}{A units vector with the measured Al content [\eqn{\mu}g/g].}
\item{Al_flag}{A logical vector indicating if the calibration range was exceeded for \code{Al}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Si}{A units vector with the measured Si content [\eqn{\mu}g/g].}
\item{Si_flag}{A logical vector indicating if the calibration range was exceeded for \code{Si}
(\code{TRUE}) or not (\code{FALSE}).}
\item{P}{A units vector with the measured P content [\eqn{\mu}g/g].}
\item{P_flag}{A logical vector indicating if the calibration range was exceeded for \code{P}
(\code{TRUE}) or not (\code{FALSE}).}
\item{S}{A units vector with the measured S content [\eqn{\mu}g/g].}
\item{S_flag}{A logical vector indicating if the calibration range was exceeded for \code{S}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Cl}{A units vector with the measured Cl content [\eqn{\mu}g/g].}
\item{Cl_flag}{A logical vector indicating if the calibration range was exceeded for \code{Cl}
(\code{TRUE}) or not (\code{FALSE}).}
\item{K}{A units vector with the measured K content [\eqn{\mu}g/g].}
\item{K_flag}{A logical vector indicating if the calibration range was exceeded for \code{K}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Ca}{A units vector with the measured Ca content [\eqn{\mu}g/g].}
\item{Ca_flag}{A logical vector indicating if the calibration range was exceeded for \code{Ca}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Ti}{A units vector with the measured Ti content [\eqn{\mu}g/g].}
\item{Ti_flag}{A logical vector indicating if the calibration range was exceeded for \code{Ti}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Cr}{A units vector with the measured Cr content [\eqn{\mu}g/g].}
\item{Cr_flag}{A logical vector indicating if the calibration range was exceeded for \code{Cr}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Mn}{A units vector with the measured Mn content [\eqn{\mu}g/g].}
\item{Mn_flag}{A logical vector indicating if the calibration range was exceeded for \code{Mn}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Fe}{A units vector with the measured Fe content [\eqn{\mu}g/g].}
\item{Fe_flag}{A logical vector indicating if the calibration range was exceeded for \code{Fe}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Cu}{A units vector with the measured Cu content [\eqn{\mu}g/g].}
\item{Cu_flag}{A logical vector indicating if the calibration range was exceeded for \code{Cu}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Zn}{A units vector with the measured Zn content [\eqn{\mu}g/g].}
\item{Zn_flag}{A logical vector indicating if the calibration range was exceeded for \code{Zn}
(\code{TRUE}) or not (\code{FALSE}).}
\item{As}{A units vector with the measured As content [\eqn{\mu}g/g].}
\item{As_flag}{A logical vector indicating if the calibration range was exceeded for \code{As}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Br}{A units vector with the measured Br content [\eqn{\mu}g/g].}
\item{Br_flag}{A logical vector indicating if the calibration range was exceeded for \code{Br}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Rb}{A units vector with the measured Rb content [\eqn{\mu}g/g].}
\item{Rb_flag}{A logical vector indicating if the calibration range was exceeded for \code{Rb}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Sr}{A units vector with the measured Sr content [\eqn{\mu}g/g].}
\item{Sr_flag}{A logical vector indicating if the calibration range was exceeded for \code{Sr}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Ba}{A units vector with the measured Ba content [\eqn{\mu}g/g].}
\item{Ba_flag}{A logical vector indicating if the calibration range was exceeded for \code{Ba}
(\code{TRUE}) or not (\code{FALSE}).}
\item{Pb}{A units vector with the measured Pb content [\eqn{\mu}g/g].}
\item{Pb_flag}{A logical vector indicating if the calibration range was exceeded for \code{Pb}
(\code{TRUE}) or not (\code{FALSE}).}
\item{C6H10O5N}{A units vector with the measured C6H10O5N content [\eqn{\mu}g/g].}
\item{C6H10O5N_flag}{A logical vector indicating if the calibration range was exceeded for \code{C6H10O5N}
(\code{TRUE}) or not (\code{FALSE}).}
}}
}
\value{
An object of class \code{xrf}. This is identical to \code{x}, but has
an additional class attribute.
}
\description{
\code{elco_new_xrf} is the internal constructor function for objects of class
\code{xrf}.
An object of class \code{xrf} is a \code{\link[tibble:tibble]{tibble::tibble()}} with a
sample in each row and a prescribed set of columns for metadata and
element contents. Objects of class \code{xrf} are intended to be created
only internally.
}
