% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_nosc.R
\name{elco_nosc}
\alias{elco_nosc}
\alias{elco_or}
\alias{elco_du}
\title{Carbon oxidation state, carbon oxidative ratio, and degree of unsaturation.}
\usage{
elco_nosc(c, h, n, o)

elco_or(c, h, n, o)

elco_du(c, h, n)
}
\arguments{
\item{c}{An object of class \code{elco} with the amount of C [mol].}

\item{h}{An object of class \code{elco} with the amount of H [mol].}

\item{n}{An object of class \code{elco} with the amount of N [mol].}

\item{o}{An object of class \code{elco} with the amount of O [mol].}
}
\value{
An object of class \code{\link[quantities:quantities]{quantities}} with the carbon oxidation state.

A numeric vector with the oxidative ratio.

A numeric vector with the oxidative ratio.
}
\description{
Functions to compute the carbon oxidation state, carbon oxidative ratio, and
degree of unsaturation with objects of class \code{\link[elco:elco_new_elco]{elco}}.
}
\section{Functions}{
\itemize{
\item \code{elco_nosc}: Computes the carbon oxidation
state using the elemental composition of C, H, N, O.

\item \code{elco_or}: Computes the carbon oxidation
state using the elemental composition of C, H, N, O.

\item \code{elco_du}: Computes the carbon degree
of unsaturation using the elemental composition of C, H, N.
}}

