% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco_math.R, R/elco_ops.R
\name{groupGeneric.elco}
\alias{groupGeneric.elco}
\alias{Math.elco}
\alias{Ops.elco}
\title{S3 Group Generic Functions}
\usage{
\method{Math}{elco}(x, ...)

\method{Ops}{elco}(e1, e2)
}
\arguments{
\item{...}{further arguments passed to methods.}
}
\description{
\code{Math}, \code{Ops} and \code{Summary} group generic methods for
\code{elco} objects (see \code{\link[base:groupGeneric]{base::groupGeneric()}} for a
comprehensive list of available methods).
}
